Refator: Project architecture for hooks and contexts
