+++ src/app/_contexts/registerFormContext.tsx
@@ -2,7 +2,7 @@
 import { createContext, useContext, useState, ReactNode } from "react";
 
 // Define the shape of the form context
-interface FormContextType {
+interface RegisterFormContextType {
   email: string;
   setEmail: (email: string) => void;
   password: string;
@@ -14,25 +14,27 @@ interface FormContextType {
 }
 
 // Create the context with default values
-const FormContext = createContext<FormContextType | undefined>(undefined);
+const RegisterFormContext = createContext<RegisterFormContextType | undefined>(
+  undefined
+);
 
 // Hook to use the context
 export const useFormContext = () => {
-  const context = useContext(FormContext);
+  const context = useContext(RegisterFormContext);
   if (!context)
     throw new Error("useFormContext must be used within a FormProvider");
   return context;
 };
 
 // Provider component to wrap around the app or pages
-export const FormProvider = ({ children }: { children: ReactNode }) => {
+export const RegisterFormProvider = ({ children }: { children: ReactNode }) => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [username, setUsername] = useState("");
   const [isOAuthUser, setIsOAuthUser] = useState(false); // New field for OAuth user
 
   return (
-    <FormContext.Provider
+    <RegisterFormContext.Provider
       value={{
         email,
         setEmail,
@@ -45,6 +47,6 @@ export const FormProvider = ({ children }: { children: ReactNode }) => {
       }}
     >
       {children}
-    </FormContext.Provider>
+    </RegisterFormContext.Provider>
   );
 };
