+++ src/app/_hooks/useConfirmPasswordVisibility.tsx
@@ -1,0 +1,15 @@
+// hooks/useConfirmPasswordVisibility.ts
+import { useState } from "react";
+
+export const useConfirmPasswordVisibility = () => {
+  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
+
+  const toggleConfirmPasswordVisibility = () => {
+    setShowConfirmPassword((prevState) => !prevState);
+  };
+
+  return {
+    showConfirmPassword,
+    toggleConfirmPasswordVisibility,
+  };
+};
