+++ src/app/_hooks/usePasswordVisibility.tsx
@@ -1,0 +1,15 @@
+// hooks/usePasswordVisibility.ts
+import { useState } from "react";
+
+export const usePasswordVisibility = () => {
+  const [showPassword, setShowPassword] = useState(false);
+
+  const togglePasswordVisibility = () => {
+    setShowPassword((prev) => !prev);
+  };
+
+  return {
+    showPassword,
+    togglePasswordVisibility,
+  };
+};
