--- src/app/main/page.tsx
@@ -1,3 +1,0 @@
-export default function MAIN() {
-  return <h1>MAIN</h1>;
-}
